refactor: timeline naming 수정
