+++ src/page/dashboard/component/Timeline/index.tsx
@@ -2,7 +2,7 @@ import { Suspense } from 'react';
 
 import DateProvider from '@/page/archiving/index/DateProvider';
 import TimeLineHeader from '@/page/archiving/index/component/TimeLine/TimeLineHeader/TimeLineHeader';
-import TimelineinDash from '@/page/dashboard/component/Timeline/Timeline/TimelineinDash';
+import Timeline from '@/page/dashboard/component/Timeline/Timeline/Timeline';
 
 const TimelineSection = () => {
   const teamId = localStorage.getItem('teamId');
@@ -11,7 +11,7 @@ const TimelineSection = () => {
     <DateProvider teamId={teamId!}>
       <TimeLineHeader />
       <Suspense>
-        <TimelineinDash />
+        <Timeline />
       </Suspense>
     </DateProvider>
   );
